resume_optimizer/requirements.txt
groq
pydantic
pdfplumber
python-docx
python-dotenv
